reorganize folders to support medata functions
