+++ python-package/basedosdados/__init__.py
@@ -12,6 +12,8 @@ from basedosdados.download.download import (
     read_sql,
     download,
     read_table,
+)
+from basedosdados.download.metadata import (
     list_datasets,
     list_dataset_tables,
     get_table_description,
@@ -19,10 +21,3 @@ from basedosdados.download.download import (
     get_table_columns,
     get_table_size,
 )
-from basedosdados.upload.table import Table
-from basedosdados.upload.storage import Storage
-from basedosdados.upload.dataset import Dataset
-import sys
-import os
-
-sys.path.append(os.getcwd() + "/python-package")
